backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
distro==1.9.0
    # via -r requirements.in
fastapi==0.115.12
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.10.0
    # via openai
markdown2==2.5.4
    # via -r requirements.in
openai==1.86.0
    # via -r requirements.in
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   uvicorn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.46.2
    # via fastapi
tabulate==0.9.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.14.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn[standard]==0.34.3
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
psycopg==3.2.12 
asyncpg